+++ tests/web_reporter_app/client/client-pass.tests.js
@@ -1,0 +1,6 @@
+/* eslint-env mocha */
+
+describe('client suite', function () {
+  it('passing test', function () {
+  });
+});