+++ tests/web_reporter_app/server/server-pass.tests.js
@@ -1,0 +1,6 @@
+/* eslint-env mocha */
+
+describe('server suite', function () {
+  it('passing test', function () {
+  });
+});