+++ package/package.js
@@ -3,7 +3,7 @@ Package.describe({
   summary: 'Run Meteor package or app tests with Mocha',
   git: 'https://github.com/meteortesting/meteor-mocha.git',
   documentation: '../README.md',
-  version: '0.4.4',
+  version: '0.5.0',
   testOnly: true,
 });
 
@@ -14,7 +14,7 @@ Package.onUse(function onUse(api) {
   ]);
 
   api.use([
-    'meteortesting:browser-tests@0.1.1'
+    'meteortesting:browser-tests@0.1.2'
   ], 'server');
 
   api.mainModule('client.js', 'client');